+++ rfq/quoting-api/config.yaml
@@ -7,3 +7,4 @@ bridges:
   421614: "0xc3c996af7f7a9245685722e73b6f18f3bb22db3d"
   11155111: "0xc3c996af7f7a9245685722e73b6f18f3bb22db3d"
 
+port: 8080
