+++ rfq/quoting-api/internal/rest/rest.go
@@ -11,8 +11,8 @@ import (
 	"github.com/synapsecns/sanguine/rfq/quoting-api/internal/bindings"
 	"github.com/synapsecns/sanguine/rfq/quoting-api/internal/config"
 	"github.com/synapsecns/sanguine/rfq/quoting-api/internal/db"
-	"github.com/synapsecns/sanguine/rfq/quoting-api/internal/db/models"
 	"github.com/synapsecns/sanguine/rfq/quoting-api/internal/rest/auth"
+	"github.com/synapsecns/sanguine/rfq/quoting-api/models"
 	omnirpcClient "github.com/synapsecns/sanguine/services/omnirpc/client"
 	"strconv"
 	"time"
