+++ rfq/quoting-api/internal/rest/rest_test.go
@@ -8,9 +8,9 @@ import (
 	"github.com/synapsecns/sanguine/rfq/quoting-api/internal/bindings"
 	"github.com/synapsecns/sanguine/rfq/quoting-api/internal/config"
 	"github.com/synapsecns/sanguine/rfq/quoting-api/internal/db"
-	"github.com/synapsecns/sanguine/rfq/quoting-api/internal/db/models"
 	"github.com/synapsecns/sanguine/rfq/quoting-api/internal/rest"
 	"github.com/synapsecns/sanguine/rfq/quoting-api/internal/testutil"
+	"github.com/synapsecns/sanguine/rfq/quoting-api/models"
 	"io"
 	"net/http"
 	"net/http/httptest"
