+++ rfq/rfq-relayer/service/relayer.go
@@ -100,7 +100,7 @@ func NewRelayer(ctx context.Context, cfg *config.Config, db db.DB, handler metri
 	}
 
 	// Create the quoter (balance management, unconfirmed bridge events, quoter API connection.
-	quoter, err := quote.NewQuoter(ctx, evmClients, cfg.Assets, common.HexToAddress(cfg.RelayerAddress), cfg.RFQURL)
+	quoter, err := quote.NewQuoter(ctx, evmClients, cfg.Assets, common.HexToAddress(cfg.RelayerAddress), cfg.RFQURL, signer)
 	if err != nil {
 		return nil, fmt.Errorf("could not create quoter: %w", err)
 	}
